+++ nni/retiarii/nn/pytorch/cell.py
@@ -11,7 +11,7 @@ import torch.nn as nn
 
 from .api import ChosenInputs, LayerChoice, InputChoice
 from .nn import ModuleList  # pylint: disable=no-name-in-module
-from .utils import generate_new_label
+from .mutation_utils import generate_new_label
 
 
 class _ListIdentity(nn.Identity):
